+++ src/api.ts
@@ -1,5 +1,5 @@
 export * as TsGenerator from "./internal/TsGenerator";
-export * as Converter from "./internal/Converter";
+export * as OpenApiTools from "./internal/OpenApiTools";
 export { FileSystem } from "./internal/FileSystem";
 export * as ResolveReference from "./internal/ResolveReference";
 export * as Validator from "./internal/Validator";
