+++ src/index.ts
@@ -1,6 +1,6 @@
 import { EOL } from "os";
 
-import { Converter, ResolveReference, TsGenerator, Validator, FileSystem } from "./api";
+import { OpenApiTools, ResolveReference, TsGenerator, Validator, FileSystem } from "./api";
 import * as Types from "./types";
 
 export interface GeneratorTemplate<T> {
@@ -11,15 +11,15 @@ export interface GeneratorTemplate<T> {
 export class CodeGenerator {
   private rootSchema: Types.OpenApi.Document;
   private resolvedReferenceDocument: Types.OpenApi.Document;
-  private parser: Converter.Parser;
+  private parser: OpenApiTools.Parser;
   constructor(private readonly entryPoint: string) {
     this.rootSchema = FileSystem.loadJsonOrYaml(entryPoint);
     this.resolvedReferenceDocument = ResolveReference.resolve(entryPoint, entryPoint, JSON.parse(JSON.stringify(this.rootSchema)));
     this.parser = this.createParser();
   }
 
-  private createParser(allowOperationIds?: string[]): Converter.Parser {
-    return new Converter.Parser(this.entryPoint, this.rootSchema, this.resolvedReferenceDocument, {
+  private createParser(allowOperationIds?: string[]): OpenApiTools.Parser {
+    return new OpenApiTools.Parser(this.entryPoint, this.rootSchema, this.resolvedReferenceDocument, {
       allowOperationIds: allowOperationIds,
     });
   }
@@ -42,12 +42,12 @@ export class CodeGenerator {
       }
       return statements;
     };
-    return [Converter.generateLeading(this.resolvedReferenceDocument), TsGenerator.generate(create)].join(EOL + EOL + EOL);
+    return [OpenApiTools.generateLeading(this.resolvedReferenceDocument), TsGenerator.generate(create)].join(EOL + EOL + EOL);
   }
 
   public generateCode<T>(generatorTemplate: GeneratorTemplate<T>): string {
     const payload = this.parser.getCodeGeneratorParamsArray();
     const create = () => TsGenerator.Utils.convertIntermediateCodes(generatorTemplate?.generator(payload, generatorTemplate.option));
-    return [Converter.generateLeading(this.resolvedReferenceDocument), TsGenerator.generate(create)].join(EOL + EOL + EOL);
+    return [OpenApiTools.generateLeading(this.resolvedReferenceDocument), TsGenerator.generate(create)].join(EOL + EOL + EOL);
   }
 }
