+++ src/internal/OpenApiTools/Comment.ts
