+++ src/internal/OpenApiTools/Guard.ts
