+++ src/internal/OpenApiTools/TypeNodeContext.ts
