+++ src/internal/OpenApiTools/components/ExternalDocumentation.ts
