+++ src/internal/OpenApiTools/components/Header.ts
