+++ src/internal/OpenApiTools/components/Headers.ts
