+++ src/internal/OpenApiTools/components/MediaType.ts
