+++ src/internal/OpenApiTools/components/Operation.ts
