+++ src/internal/OpenApiTools/components/Parameter.ts
