+++ src/internal/OpenApiTools/components/PathItem.ts
