+++ src/internal/OpenApiTools/components/PathItems.ts
