+++ src/internal/OpenApiTools/components/Reference.ts
