+++ src/internal/OpenApiTools/components/Schema.ts
