+++ src/internal/OpenApiTools/components/SecuritySchema.ts
