+++ src/internal/OpenApiTools/components/SecuritySchemas.ts
