+++ src/internal/OpenApiTools/paths/index.ts
