+++ src/internal/OpenApiTools/store/Definition.ts
