+++ src/internal/OpenApiTools/store/Operation.ts
