+++ src/internal/OpenApiTools/store/Store.ts
