+++ src/internal/OpenApiTools/store/index.ts
