+++ src/internal/OpenApiTools/store/structure/DataStructure.ts
