+++ src/internal/OpenApiTools/store/structure/InterfaceNode.ts
