+++ src/internal/OpenApiTools/store/structure/NamespaceTree.ts
