+++ src/internal/OpenApiTools/store/structure/TypeAliasNode.ts
