+++ src/internal/OpenApiTools/store/structure/index.ts
