+++ src/internal/OpenApiTools/toTypeNode.ts
